deployment to ms
git-svn-id: svn+ssh://zene.sk/home/tomik/svn/akimot/trunk@97 6bcf84ba-f5f3-4bf5-b896-c696636b5bce

+++ board.cpp
@@ -1054,7 +1054,7 @@ void Board::commitMove()
 
   toMove_ = OPP(toMove_);
   toMoveIndex_ = 1 - toMoveIndex_;
-  assert(toMoveIndex_ == PLAYER_TO_INDEX(toMove_));
+  assert(toMoveIndex_ == uint(PLAYER_TO_INDEX(toMove_)));
   stepCount_ = 0;
 
   //preMoveSignature of the next move is the actual signature
